requirements.txt
streamlit
pymupdf
openai
Pillow